Frameworks: Gesture Lockscreen (1/2)
This adds the ability to us a hand drawn gesture as
a means to unlock the device.

In the case that the gesture is forgotton or not
being recognized correctly, the fallback will be the
same as when using patterns.

Androidx:
Some changes needed for 5.0

@alviteri: crDroid'fy

(cherry picked from commit a1a9db0ce13da0ede769e76ae4332d06c9a556d7)