Frameworks: Be sure heads up view is attached before removing it
We did not have it in our tests but some user ran into it. Thing is if the
system is disabling notifications for whatever reason and till this time the
view was never attached and the system tries to detach the view on windowmanager
we run into an illegal argument exception.

Fix it with simply checking if the view was attached before removing it.

Change-Id: I0800beed3553815cc529087052138345773b32e4

(cherry picked from commit 05e605ab3758fe97d41c5c62056a9cb1c4f6f75c)

(cherry picked from commit 69143a974669e3b76b956a96a972e72eb40982ea)